services/document/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3
opencv-python==4.8.1.78

